Fix deadlock on app start due incorrect thread usage of lifecycle methods
See https://issuetracker.google.com/issues/136885494
